requirements.txt
beautifulsoup4
# make numpy less than 2.3
numpy<2.3
pandas
tqdm
joblib
lxml  # For better XML parsing performance (if not already installed with xml.etree)
unidecode
streamlit
plotly
scikit-learn
pyarrow
fastapi
uvicorn
sentence-transformers
crossrefapi
pyyaml
requests
google-generativeai
pyyaml
torch
dropbox
google-genai
python-doi
pytorch_lightning
pynndescent
umap-learn
dill
faiss-cpu
rank_bm25
